+++ BE/PoroDev.Runtime/Program.cs
@@ -1,6 +1,4 @@
 using PoroDev.Common.MassTransit;
-using PoroDev.Runtime.Extensions;
-using PoroDev.Runtime.Extensions.Contracts;
 using PoroDev.Runtime.Mapper;
 using PoroDev.Runtime.Services;
 using PoroDev.Runtime.Services.Contracts;
@@ -18,6 +16,7 @@ builder.Services.AddSwaggerGen();
 builder.Services.AddScoped<IRuntimeService, RuntimeService>();
 builder.Services.AddScoped<IDockerImageService, DockerImageService>();
 builder.Services.AddScoped<IZipManipulator, ZipManipulator>();
+builder.Services.AddScoped<IRuntimeHelper, RuntimeHelper>();
 var app = builder.Build();
 
 // Configure the HTTP request pipeline.
