+++ BE/PoroDev.Runtime/Services/Contracts/IRuntimeHelper.cs
@@ -1,0 +1,10 @@
+﻿using PoroDev.Common.Contracts;
+using PoroDev.Common.Models.RuntimeModels.Data;
+
+namespace PoroDev.Runtime.Services.Contracts
+{
+    public interface IRuntimeHelper
+    {
+        CommunicationModel<RuntimeData> InitializeAndExtract();
+    }
+}
