+++ BE/PoroDev.Runtime/Services/Contracts/IZipManipulator.cs
@@ -1,6 +1,6 @@
 ﻿using PoroDev.Common.Exceptions;
 
-namespace PoroDev.Runtime.Extensions.Contracts
+namespace PoroDev.Runtime.Services.Contracts
 {
     public interface IZipManipulator
     {
